Safari-609.1.11.1/LayoutTests/http/tests/referrer-policy-script/same-origin/cross-origin-http.https-expected.txt

Tests the behavior of same-origin referrer policy when cross origin from HTTPS to HTTP.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is ""
PASS successfullyParsed is true

TEST COMPLETE
